modularized everything, vector is fixed atm bc I haven't fixed the math yet
