--- src/animal.py
@@ -1,10 +1,0 @@
-class Animal:
-
-    def __init__(self, direction):
-        self.speed = 10      // # of times animal can move/60 frames
-        self.direct = direction
-     
-    def _move(self):
-        pass        
-        
-
