+++ src/player.py
@@ -1,0 +1,8 @@
+from pathlib import Path
+import pygame
+
+class Player:
+    def __init__(self, image: str, size: tuple, location: tuple = (0, 0)):
+        self.img = pygame.transform.scale(pygame.image.load(image), size)
+        self.rect = self.img.get_rect()
+        self.rect = self.rect.move(location)
